requirements.txt
langchain
langchain-community
langchain-experimental
openai
ollama
duckduckgo-search
python-dotenv
streamlit
chromadb
